+++ rail/creation/__init__.py
@@ -1,4 +1,4 @@
 """Modules to create synthetic data"""
 
-from rail.creation.creator import Creator, Modeler, PosteriorCalculator
 from rail.creation.degrader import Degrader
+from rail.creation.engine import Creator, Modeler, PosteriorCalculator
