--- rail/creation/engines/__init__.py
@@ -1,2 +1,0 @@
-"""Engines that create synthetic samples of photometric data"""
-from rail.creation.engines.flowCreator import FlowCreator, FlowModeler, FlowPosterior
